+++ gargoyle/builtins.py
@@ -5,6 +5,8 @@ gargoyle.builtins
 :copyright: (c) 2010 DISQUS.
 :license: Apache License 2.0, see LICENSE for more details.
 """
+from __future__ import absolute_import, division, print_function, unicode_literals
+
 import socket
 import struct