+++ gargoyle/constants.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import, division, print_function, unicode_literals
+
 DISABLED = 1
 SELECTIVE = 2
 GLOBAL = 3