+++ gargoyle/manager.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import, division, print_function, unicode_literals
+
 from django.conf import settings
 from django.core.cache import caches
 from django.http import HttpRequest