+++ gargoyle/nexus_modules.py
@@ -5,6 +5,8 @@ gargoyle.nexus_modules
 :copyright: (c) 2010 DISQUS.
 :license: Apache License 2.0, see LICENSE for more details.
 """
+from __future__ import absolute_import, division, print_function, unicode_literals
+
 import logging
 from functools import wraps