+++ gargoyle/templatetags/gargoyle_helpers.py
@@ -5,6 +5,8 @@ gargoyle.templatetags.gargoyle_helpers
 :copyright: (c) 2010 DISQUS.
 :license: Apache License 2.0, see LICENSE for more details.
 """
+from __future__ import absolute_import, division, print_function, unicode_literals
+
 from django import template
 
 register = template.Library()