+++ runtests.py
@@ -57,6 +57,8 @@ def run_isort():
     print('Running isort check')
     return subprocess.call([
         'isort', '--recursive', '--check-only', '--diff',
+        '-a', 'from __future__ import absolute_import, division',
+        '-a', 'from __future__ import print_function, unicode_literals',
         'gargoyle', 'tests'
     ])