+++ tests/manage.py
@@ -2,6 +2,8 @@
 """
 Used for running this install of gargoyle locally
 """
+from __future__ import absolute_import, division, print_function, unicode_literals
+
 import os
 import sys