+++ tests/settings/base.py
@@ -1,4 +1,6 @@
 # -*- coding:utf-8 -*-
+from __future__ import absolute_import, division, print_function, unicode_literals
+
 import os
 
 BASE_DIR = os.path.dirname(os.path.dirname(__file__))