+++ tests/settings/test.py
@@ -1,6 +1,8 @@
 """
 Used for the test suite run.
 """
+from __future__ import absolute_import, division, print_function, unicode_literals
+
 from copy import deepcopy
 
 from .base import *  # noqa