+++ tests/testapp/management/commands/runserver.py
@@ -2,6 +2,8 @@
 Exists to ensure you can always log in with admin/password when running in
 'dev' mode - see tests/README.rst
 """
+from __future__ import absolute_import, division, print_function, unicode_literals
+
 from django.contrib.auth.models import User
 from django.contrib.staticfiles.management.commands.runserver import Command as BaseCommand
 from django.core.management import call_command