+++ tests/testapp/test_api.py
@@ -2,6 +2,8 @@
 :copyright: (c) 2010 DISQUS.
 :license: Apache License 2.0, see LICENSE for more details.
 """
+from __future__ import absolute_import, division, print_function, unicode_literals
+
 import datetime
 
 from django.contrib.auth.models import AnonymousUser, User