+++ tests/testapp/test_helpers.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import, division, print_function, unicode_literals
+
 from django.contrib.auth.models import AnonymousUser, User
 from django.test import TestCase