+++ tests/testapp/test_template_tags.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import, division, print_function, unicode_literals
+
 from django.contrib.auth.models import User
 from django.http import HttpRequest
 from django.template import Context, Template, TemplateSyntaxError