+++ tests/urls.py
@@ -2,6 +2,8 @@
 :copyright: (c) 2010 DISQUS.
 :license: Apache License 2.0, see LICENSE for more details.
 """
+from __future__ import absolute_import, division, print_function, unicode_literals
+
 import nexus
 from django.conf.urls import include, url
 from django.contrib import admin