+++ HISTORY.rst
@@ -10,6 +10,8 @@ Pending Release
 * New release notes here
 * Removed the South Migrations, since South doesn't support Django 1.7+, and
   Gargoyle only supports Django 1.8+.
+* Added all ``__future__`` imports to all files for Python 2.7/3
+  compatibility.
 
 1.2.1 (2016-02-25)
 ------------------